+++ desktop/src/main/java/com/onecat/burst/utils/InputChecks.java
@@ -1,0 +1,29 @@
+package com.onecat.burst.utils;
+
+public class InputChecks {
+
+	public static boolean isInteger(String str) {
+		if (str == null || str.isEmpty()) {
+			return false;
+		}
+		try {
+			Integer.parseInt(str);
+			return true;
+		} catch (NumberFormatException e) {
+			return false;
+		}
+	}
+
+	public static boolean isFloat(String str) {
+		if (str == null || str.isEmpty()) {
+			return false;
+		}
+		try {
+			Float.parseFloat(str);
+			return true;
+		} catch (NumberFormatException e) {
+			return false;
+		}
+	}
+
+}