+++ docs/changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## Camera, PFX loading, Settings
+**Aug 30, 2026**
+- Made gizmo a bit thinner
+- Added Skin Composer assets and styles for PopColorPicker
+- Added a controllable orbiting camera
+- Particle presets can now be loaded and get rendered, yippee
+- Split EditorUI elements into separate classes
+- Implemented EditorPanel with editor settings
+- Implemented TopBar and ControllersPanel UI elements (WIP)
+- Editor now uses Preferences
+
 ## Acting busy
 **Aug 28, 2026**
 - Added SkinComposer project to the repo